requirements.txt
aiogram==3.10.0
aiohttp==3.9.3
python-dotenv==1.0.0
gigachat==0.1.13
